Everything is normed
