+++ srcs/exec/exec.h
@@ -6,7 +6,7 @@
 /*   By: tmatis <tmatis@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/04 14:12:28 by tmatis            #+#    #+#             */
-/*   Updated: 2021/05/22 19:40:48 by tmatis           ###   ########.fr       */
+/*   Updated: 2021/05/23 15:15:03 by tmatis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,5 +33,10 @@ void	execution_error_status(char *cmd, char *error, int status,
 int		handle_dup2_error(t_tube *tube_list, int forks_running, int fork_n);
 void	sig_handler(int signal_no);
 int		child_process(t_child_data child_data);
+void	*return_and_free(char **envp, int i);
+void	free_until(char **table, int max);
+char	**build_env(t_list *env_var);
+char	**build_argv(char *name, t_list *args);
+int		build_argc(char **argv);
 
 #endif
